Address feedback and lint
