--- packages/contracts/src/utils/utils.ts
@@ -1,7 +1,0 @@
-export const utils = {
-    consoleLog(message: string): void {
-        /* tslint:disable */
-        console.log(message);
-        /* tslint:enable */
-    },
-};
