+++ packages/contracts/test/exchange/wrapper.ts
@@ -741,4 +741,4 @@ describe('Exchange', () => {
             });
         });
     });
-});
+}); // tslint:disable-line:max-file-line-count
